backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
vnstock==3.2.1
pandas==2.2.3
openpyxl==3.1.20
beautifulsoup4==4.12.3
requests==2.32.3
python-dotenv==1.0.1
python-multipart==0.0.19
pydantic==2.10.4
pdfplumber==0.11.0
groq==0.5.0
google-generativeai==0.8.3
